backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
google-generativeai
langchain-groq
pydantic
sqlalchemy
python-dotenv
pandas
psycopg2-binary
